Rename fixture project files
