+++ packages/tests/src/fixtures/project/main1.js
@@ -2,8 +2,10 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import fn from './src/file0000.js';
-import fn2 from './workspaces/app/file0001.js';
+// project/main1.js
+
+import fn from './src/srcFile0.js';
+import fn2 from './workspaces/app/workspaceFile1.js';
 
 // Add a third party dependency.
 import * as chalk from 'chalk';
