+++ packages/tests/src/fixtures/project/src/srcFile0.js
@@ -2,7 +2,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import fn2 from '../workspaces/app/file0000.js';
+// project/src/srcFile0.js
+
+import fn2 from '../workspaces/app/workspaceFile0.js';
 
 const fn = async function () {
     await fn2();
