+++ packages/tests/src/fixtures/project/workspaces/app/workspaceFile1.js
@@ -2,7 +2,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import fn1 from './file0000.js';
+// project/workspaces/app/workspaceFile1.js
+
+import fn1 from './workspaceFile0.js';
 
 const fn = async function () {
     await fn1();
